+++ e2e/pages/administration/apiToken.po/constants.ts
@@ -1,10 +1,10 @@
 import { by, element } from 'protractor';
 import { Autocomplete } from '../../../elements/autocomplete.element';
 
-export const baseUrl = '#/administration/project/importToken';
+export const baseUrl = '#/administration/project/apiToken';
 
 export const elements = {
-    uniqueElement: element(by.css('#import-token-administration.active')),
+    uniqueElement: element(by.css('#api-token-administration.active')),
     projectSelector: new Autocomplete(by.id('project-selector')),
     generateToken: element(by.id('generateTocken')),
     tokenValue: element(by.id('tokenValue'))